requirements.txt
streamlit
streamlit-option-menu
snowflake-connector-python
pandas~=1.4.2
Pillow~=9.0.1
requests~=2.27.1
streamlit-aggrid==0.3.3
altair~=4.2.0
graphviz~=0.20.1
wordcloud
streamlit-elements
streamlit-login-page==0.0.7
